Add files via upload
These scripts here demonstrate regression modelling using ctrees and random forests, and how to downsample a dataset before growing a random forest. The first script applies ctrees (conditional inference trees) and random forests to a historical dataset to analyse variation in language use. For this example I am using a dataset of 18th century Scots and English words present in a corpus of texts. The second script refers to a subset of this dataset - here we are interested in only authors with political sympathies (coded as Pro-union or Anti-Union), and shows how to downsample this subset to reduce class imbalance in the dependent variable (Scots_English - i.e. Scots words and English words), before growing a random forest from this data.